Self activate the VC Editor
The latest version of VC and Popup Maker don't work with the VC Editor on the popup post type.
I'm running some filters on the post type itself at runtime which should always work if they don't change the filter of course. Hopefully, they change this less than they change everything else, haha. 

One thing that is missing is outputting the custom css from VC's editor that VC stores as option _wpb_custom_css. I'm working on a custom theme right now, so when I write a function for this myself I will make another push after testing it on Popup Maker. 

Last... I didn't look at how you are determining the Popup Maker post types. I run my own function for determining the post type. For that I called that function "Popmake_GetCurrentPostType()". You might replace that with the function you are using to reduce code or use my function (which has come in handy from a dev that gave it to me) for your function and still reduce the code. 

It's up to you.