Reverted back to use the init action
It seems this was changed again by mistake. The changes to how we
register actions manually was done so that we could initialize this
later than plugins_loaded so that theme's could unhook it if need be.

When using plugins_loaded, the manual registration of actions was
occurring before Ninja Forms core did, and were overwritten. So they
did not appear any longer. Updating to init resolved this issue.