Refactored internal value parsing of settings
Previously it did not parse in default values on the way in and out for
settings. Now it does both using a new fill_missing_defaults method.
This method excludes checkbox fields to prevent rechecking unchecked
fields.

This also fixes an issue with us being unable to set std = true
for checkbox fields.

Related to #592