Refactor activation & init to check PHP version
Newly added activation callback that will short circuit the plugins
activation if minimum PHP & WP requirements are not met. This will
essentially throw a notice to what is required and deactivate the
plugin. This should only be needed on fresh installs to sites running
below WP v5.1.

Just in case added an additional check in the plugin init to abort in
case the plugin is active already on a server with older version of PHP.
The main plugin file will of course need to stay clear of code that
requires later versions of PHP until the bulk of our users have migrated
to latest versions of PHP. It will also show an admin notice to indicate
something isn't correct.

Related to #721