Create extend with other plugins renderer
Since the empty elseif was caught in scans that needed to be resolved
for the code quality changes, I went ahead and quickly created the
renderer instead of deleting that branch of the if/else. Still need to
add plugins to recommend but at least it's off the code quality list.

Issue #718