Improve getting uploads directory for generating and enqueueing cache
We needed to add in a few checks to ensure the uploads directory system
was working as well as making sure no errors were coming out of it. So,
went ahead and rewrote some of the functions by creating 3 new helpers
as well as moved the main calls to each into the variable in the assets
and assetcache classed and tested against those instead of rerunning
the functions again as it currently did.