Corrects an issue with getting uploads directory
When I reviewed the docs for the wp_upload_dir() functions, it made
it seem like the 'error' key would only be present if an error occurred
and would be false if unknown reason or a string with a reason. Turns
out, that was an incorrect understanding of the function. The returned
array always has an 'error' key. This misunderstanding was causing the
cache_url to remain false causing a site to not load the cache files.