Add overrides for assets folder
We can set the whole repo to esnext so our newer JS files, such as for the block editor, can lint using esnext. However, this then shows errors in our assets JS that we cannot fix, such as using let instead of var. So, we use overrides to prevent these.

Issue #822