Ensure GF AJAX submissions are only counted once
Gravity Forms doesn't use the built-in admin_ajax system for
submitting via AJAX. Instead, they submit the form data as POST data
back to the page the form is on. Luckily, they set the gform_ajax key
when submitting via AJAX so we can bail early in the PHP handler if
that is present.